[Cleanup] Remove hyperparams/config_definitions.py. The classes are only for deprecated models. Move the usages to individual legacy models and we will remove them finally.
PiperOrigin-RevId: 408997960
