+++ official/projects/basnet/configs/basnet.py
@@ -16,11 +16,10 @@
 import dataclasses
 import os
 from typing import List, Optional, Union
-
+from official.core import config_definitions as cfg
 from official.core import exp_factory
 from official.modeling import hyperparams
 from official.modeling import optimization
-from official.modeling.hyperparams import config_definitions as cfg
 from official.vision.beta.configs import common
