+++ official/projects/basnet/configs/basnet_test.py
@@ -18,8 +18,8 @@
 from absl.testing import parameterized
 import tensorflow as tf
 
+from official.core import config_definitions as cfg
 from official.core import exp_factory
-from official.modeling.hyperparams import config_definitions as cfg
 from official.projects.basnet.configs import basnet as exp_cfg
