+++ official/projects/edgetpu/vision/configs/semantic_segmentation_config.py
@@ -21,11 +21,10 @@ deeplab v3 segmentation head.
 import dataclasses
 import os
 from typing import Optional
-
+from official.core import config_definitions as cfg
 from official.core import exp_factory
 from official.modeling import hyperparams
 from official.modeling import optimization
-from official.modeling.hyperparams import config_definitions as cfg
 from official.vision.beta.configs import backbones
 from official.vision.beta.configs import common
 from official.vision.beta.configs import decoders
