+++ official/projects/volumetric_models/configs/semantic_segmentation_3d.py
@@ -17,10 +17,10 @@
 
 import dataclasses
 from typing import List, Optional, Union
+from official.core import config_definitions as cfg
 from official.core import exp_factory
 from official.modeling import hyperparams
 from official.modeling import optimization
-from official.modeling.hyperparams import config_definitions as cfg
 from official.projects.volumetric_models.configs import backbones
 from official.projects.volumetric_models.configs import decoders
 from official.vision.beta.configs import common
