+++ official/vision/beta/projects/centernet/configs/centernet.py
@@ -17,11 +17,10 @@
 import dataclasses
 import os
 from typing import List, Optional, Tuple
-
+from official.core import config_definitions as cfg
 from official.core import exp_factory
 from official.modeling import hyperparams
 from official.modeling import optimization
-from official.modeling.hyperparams import config_definitions as cfg
 from official.vision.beta.configs import common
 from official.vision.beta.projects.centernet.configs import backbones
