+++ official/vision/beta/projects/panoptic_maskrcnn/serving/panoptic_segmentation.py
@@ -18,7 +18,7 @@ from typing import List
 
 import tensorflow as tf
 
-from official.modeling.hyperparams import config_definitions as cfg
+from official.core import config_definitions as cfg
 from official.vision.beta.projects.panoptic_maskrcnn.modeling import panoptic_maskrcnn_model
 from official.vision.beta.serving import detection
