+++ official/vision/image_classification/configs/configs.py
@@ -52,7 +52,7 @@ class EfficientNetImageNetConfig(base_configs.ExperimentConfig):
           enable_checkpoint_and_export=True, enable_tensorboard=True),
       metrics=['accuracy', 'top_5'],
       time_history=base_configs.TimeHistoryConfig(log_steps=100),
-      tensorboard=base_configs.TensorboardConfig(
+      tensorboard=base_configs.TensorBoardConfig(
           track_lr=True, write_model_weights=False),
       set_epoch_loop=False)
   evaluation: base_configs.EvalConfig = base_configs.EvalConfig(
@@ -84,7 +84,7 @@ class ResNetImagenetConfig(base_configs.ExperimentConfig):
           enable_checkpoint_and_export=True, enable_tensorboard=True),
       metrics=['accuracy', 'top_5'],
       time_history=base_configs.TimeHistoryConfig(log_steps=100),
-      tensorboard=base_configs.TensorboardConfig(
+      tensorboard=base_configs.TensorBoardConfig(
           track_lr=True, write_model_weights=False),
       set_epoch_loop=False)
   evaluation: base_configs.EvalConfig = base_configs.EvalConfig(
